transport/libfabric: Use FI_ORDER_SAS with staged atomics
Staged atomics require FI_ORDER_SAS for correctness when the user
invokes nvshmem_fence(). Example:

nvshmem_atomic_set(1)
nvshmem_fence()
nvshmem_atomic_set(5)

nvshmem_fence() guarantees that the value of 5 will always be the result
of the program. The implementation of nvshmem_fence() for libfabric is
to block until there are TX completions available for every submitted
op. The libfabric transport is currently using completion semantic
FI_TRANSMIT_COMPLETE, defined by the man pages for reliable endpoints as:

Indicates that a completion should be generated when the operation has
been delivered to the peer endpoint. A completion guarantees that the
operation is no longer dependent on the fabric or local resources.
The state of the operation at the peer endpoint is not defined.

This means that the staged_atomics fi_send has arrived to the bounce
buffer on the RX side, and a completion is in EFA's internal IBV_CQ.
The libfabric API does not guarantee that internal completions will
be reported in the order that they were received unless FI_ORDER_SAS
is set.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>