host/topo: Fix indexing bug causing wrong GPU->NIC assignments
When a NIC that has more than 1 GPU is attempting to be rebalanced, the
IF check ((*pairs_iter).pe_idx != nic_id) confirms that the distance
between NIC/GPU is for the GPU that is trying to be rebalanced.
This indexing bug causes GPU-NIC rebalancing to happen based on
distance of the wrong GPU->NIC pair. If there are 8 PE's per node,
and 2 NICS per PE, and PE 1 is attempting to be rebalanced, this
bug would cause the distance between PE 0 -> NIC X to be used
instead of PE 1 -> NIC X.

Co-authored-by: Danylo Vashchilenko <vdanylo@amazon.com>
Signed-off-by: Seth Zegelstein <szegel@amazon.com>

SIM: https://sim.amazon.com/issues/Subspace-3169
cr: https://code.amazon.com/reviews/CR-204593994