transport/libfabric: Make gdr put-signal send 32 byte signal packet
The EFA Provider has a 32 byte inline send data size. Previously, the put-signal
implementation for gdrcopy emulated atomics called the already existing atomic
functions in order to optimize code re-use. The size of an atomic operation wire
data packet is 72 bytes, which is larger than the EFA provider's 32 byte inline
send size. This change modifies the wire data for the signal part of the put-signal
operation to be 32 bytes, at the expense of code reuse.  When the signal wire data
packet is recived by the RX side, the RX side will re-format that data in-place to
re-use the existing emulated gdrcopy atomic functions.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>