transport/libfabric: Make fence() correct with staged_atomics
The libfabric transport's fence() implementation for staged_atomics
depended on completion ordering (comp_order) which was not requested,
and has been deprecated from the libfabric API. The fence()/quiet()
implementation would block until all TX completions were received for
every submitted operation, but that does not guarantee that the atomic
has been applied at the peer. For the following program:

atomic(5)
fence()
atomic(8)

There is a race condition that existed where the fence finishes but the
first atomic has not been processed by the RX side.  The TX side sends
the second atomic, and both of them are sitting in the devices RX
completion queue at the same time. The libfabric API makes no guarantees
on completion ordering so atomic(8) could happen before atomic(5).

This patch fixes this race by adding a 0 byte write w/imm ack for every
staged_atomic operation. When the TX side submits an atomic, they
increase the submitted ops by 2, and the fence()/quiet() will not
complete until the ack comes back from the reciver saying that the
atomic was applied.

This patch also adds a 4 bit header to the immediate data allowing the
immediate data to be used for more than just put_signal_unordered.

For a 0 byte write w/imm, the efa device currently requires valid
paramters for local buffer, local buffer desc, rptr, and rkey even
though no data is being transmitted.  In order to meet these
requirements, a registered write target variable was added to libfabric
state, and its address/rkey are exchanged during bootstrap.  I attempted
to re-use the same atomic that had the operation performed on it, but
this was not possible because the nvshmemi_state global is not exposed
to the transport libraries.

This change is expected to make fence()/quiet() calls more expensive for
libfabric providers that use staged_atomics, but is required for
correctness.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>

SIM: https://i.amazon.com/Subspace-3174
cr: https://code.amazon.com/reviews/CR-218050813