transport/libfabric: Only 1 put-signal seq_cntr per EP
Previously, there was 1 put-signal seq_cntr per EP per Peer. This was
done to make the rate of the sequence counters go up slower to prevent
possible overflow collision race conditions.  With 28 bits, the max
value of the sequence counter is 268 million, and with a packet rate of
~2 million packets per second, the sequence counter is not likely to
overflow anyways.  This change reduces the number of sequence counters
to be 1 sequence counter per EP (proxy or host thread) which should be
nicer on cache.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>

cr: https://code.amazon.com/reviews/CR-218050813