device/col: Fix ordering race condition in barrier
The barrier algorithm uses signal sets to implement NVSHMEM barrier.
This behavior has a requirement on ordering b/c if the messages arrive
out of order, the program could deadlock waiting on a value to be set
that will never arrive. This commit adds a fence() to the barrier
algorithm making the ordering requirement explicit, and hence removing
the race.

Thank you to Seth Howell for help with debugging!

Signed-off-by: Seth Zegelstein <szegel@amazon.com>
Co-authored-by: Seth Howell <sethh@nvidia.com>